requirements.txt
flask>=3.0
pyjwt>=2.8
python-dotenv>=1.0
werkzeug>=3.0
